requirements.txt
pip==23.3.2
python-telegram-bot==20.7
requests==2.31.0
telethon==1.33.1
pyrogram==2.0.106